25304번 영수증

+++ src/baekjoon/impl/No25304.java
@@ -9,13 +9,13 @@ public class No25304 {
 
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
-        int x = Integer.parseInt(br.readLine());
+        long x = Long.parseLong(br.readLine());
         int n = Integer.parseInt(br.readLine());
 
-        int result = 0;
+        long result = 0;
         for (int i = 0; i < n; i++) {
             StringTokenizer st = new StringTokenizer(br.readLine());
-            result += Integer.parseInt(st.nextToken()) * Integer.parseInt(st.nextToken());
+            result += Long.parseLong(st.nextToken()) * Long.parseLong(st.nextToken());
         }
 
         if (result == x) {
